Add WebDriver interface
This commit utilizes the WebDriver interface which will allow a
much easier way to switch between different browsers for testing.

Adding global variables for the browser and implimenting the WebDriver
interface we can apply an if statement to determine which browser
is being tested and the WebDriverManager can setup the corresponding
driver appropriately.